:bug:fix: student_number overflow 해결

+++ src/main/java/com/be/recheckbe/domain/auth/dto/RegisterRequest.java
@@ -19,7 +19,7 @@ public class RegisterRequest {
 
   @NotBlank private String phoneNumber;
 
-  @NotNull private Integer studentNumber;
+  @NotNull private Long studentNumber;
 
   @NotNull private Long departmentId;
 }
